+++ src/app/helpers.ts
@@ -8,7 +8,7 @@ import { StructureQuery } from 'molstar/lib/mol-model/structure/query/query';
 import { BuiltInTrajectoryFormat } from 'molstar/lib/mol-plugin-state/formats/trajectory';
 import { StructureRef } from 'molstar/lib/mol-plugin-state/manager/structure/hierarchy-state';
 import { StateTransforms } from 'molstar/lib/mol-plugin-state/transforms';
-import { CreateVolumeStreamingInfo } from 'molstar/lib/mol-plugin/behavior/dynamic/volume-streaming/transformers';
+import { CreateVolumeStreamingInfo, InitVolumeStreaming } from 'molstar/lib/mol-plugin/behavior/dynamic/volume-streaming/transformers';
 import { PluginCommands } from 'molstar/lib/mol-plugin/commands';
 import { PluginConfigItem } from 'molstar/lib/mol-plugin/config';
 import { PluginContext } from 'molstar/lib/mol-plugin/context';
@@ -39,9 +39,15 @@ export interface LoadParams {
 }
 
 export interface MapParams {
+    /** Volume streaming view type (`'off' | 'box' | 'selection-box' | 'camera-target' | 'cell' | 'auto'`) */
+    defaultView?: PDBeVolumes.InitVolumeStreamingProps['defaultView'],
+    /** Style for EM map */
     'em'?: MapStyle,
+    /** Style for X-ray 2Fo-Fc map */
     '2fo-fc'?: MapStyle,
+    /** Style for X-ray Fo-Fc(+ve) map */
     'fo-fc(+ve)'?: MapStyle,
+    /** Style for X-ray Fo-Fc(-ve) map */
     'fo-fc(-ve)'?: MapStyle,
 }
 interface MapStyle {
@@ -51,10 +57,15 @@ interface MapStyle {
 
 
 export namespace PDBeVolumes {
+    export type InitVolumeStreamingProps = ReturnType<(typeof InitVolumeStreaming)['createDefaultParams']>;
 
-    export function mapParams(defaultParams: any, mapParams?: MapParams, ref?: string | number) {
+    export function mapParams(defaultParams: InitVolumeStreamingProps, mapParams?: MapParams): InitVolumeStreamingProps {
         const pdbeParams = { ...defaultParams };
-        pdbeParams.options.behaviorRef = 'volume-streaming' + '' + Math.floor(Math.random() * Math.floor(100));
+
+        if (mapParams?.defaultView) {
+            pdbeParams.defaultView = mapParams.defaultView;
+        }
+        pdbeParams.options.behaviorRef = `volume-streaming${Math.floor(Math.random() * 100)}`;
         pdbeParams.options.emContourProvider = 'pdbe';
         pdbeParams.options.serverUrl = 'https://www.ebi.ac.uk/pdbe/volume-server';
         const MAIN_MAP_DEFAULTS: MapStyle = { opacity: 0.49, wireframe: false };